# backend/services/account-service/docs/swagger.yaml
basePath: /
definitions:
...
  github_com_fintrack_account-service_internal_infrastructure_entrypoints_handlers_card_dto.CardResponse:
    properties:
      account_id:
        type: string
      card_brand:
        type: string
      card_type:
        type: string
      closing_date:
        type: string
      created_at:
        type: string
      credit_limit:
        description: Credit card specific fields
        type: number
      due_date:
        type: string
      expiration_month:
        type: integer
      expiration_year:
        type: integer
      holder_name:
        type: string
      id:
        type: string
      is_default:
        type: boolean
      last_four_digits:
        type: string
      masked_number:
        type: string
      nickname:
        type: string
      status:
        type: string
      updated_at:
        type: string
    type: object
  github_com_fintrack_account-service_internal_infrastructure_entrypoints_handlers_card_dto.CreateCardRequest:
    properties:
      account_id:
        type: string
      card_brand:
        type: string
      card_type:
        description: '"credit" or "debit"'
        type: string
      closing_date:
        type: string
      credit_limit:
        description: Credit card specific fields
        minimum: 0
        type: number
      due_date:
        type: string
      encrypted_number:
        description: Security fields (encrypted data)
        type: string
      expiration_month:
        maximum: 12
        minimum: 1
        type: integer
      expiration_year:
        type: integer
      holder_name:
        maxLength: 100
        minLength: 2
        type: string
      is_default:
        type: boolean
      key_fingerprint:
        type: string
      last_four_digits:
        type: string
      masked_number:
        type: string
      nickname:
        maxLength: 50
        type: string
    required:
    - account_id
    - card_brand
    - card_type
    - encrypted_number
    - expiration_month
    - expiration_year
    - holder_name
    - key_fingerprint
    - last_four_digits
    - masked_number
    type: object
  github_com_fintrack_account-service_internal_infrastructure_entrypoints_handlers_card_dto.PaginatedCardResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/github_com_fintrack_account-service_internal_infrastructure_entrypoints_handlers_card_dto.CardResponse'
        type: array
      pagination:
        $ref: '#/definitions/github_com_fintrack_account-service_internal_infrastructure_entrypoints_handlers_card_dto.PaginationMeta'
    type: object
  github_com_fintrack_account-service_internal_infrastructure_entrypoints_handlers_card_dto.PaginationMeta:
    properties:
      current_page:
        type: integer
      page_size:
        type: integer
      total_items:
        type: integer
      total_pages:
        type: integer
    type: object
  github_com_fintrack_account-service_internal_infrastructure_entrypoints_handlers_card_dto.UpdateCardRequest:
    properties:
      expiration_month:
        maximum: 12
        minimum: 1
        type: integer
      expiration_year:
        minimum: 2024
        type: integer
      holder_name:
        maxLength: 100
        minLength: 2
        type: string
      is_default:
        type: boolean
      nickname:
        maxLength: 50
        type: string
    type: object
host: localhost:8082
info:
  contact:
    email: support@fintrack.com
    name: API Support
    url: http://www.fintrack.com/support
  description: Account management service for FinTrack application
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: FinTrack Account Service API
  version: "1.0"
paths:
...
  /api/accounts/{id}/cards:
    get:
      consumes:
      - application/json
      description: Retrieve all cards associated with an account
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Page size
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Cards retrieved successfully
          schema:
            $ref: '#/definitions/github_com_fintrack_account-service_internal_infrastructure_entrypoints_handlers_card_dto.PaginatedCardResponse'
        "400":
          description: Invalid request parameters
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Account not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get cards by account
      tags:
      - Cards
    post:
      consumes:
      - application/json
      description: Create a new card associated with an account
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: string
      - description: Card creation data
        in: body
        name: card
        required: true
        schema:
          $ref: '#/definitions/github_com_fintrack_account-service_internal_infrastructure_entrypoints_handlers_card_dto.CreateCardRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Card created successfully
          schema:
            $ref: '#/definitions/github_com_fintrack_account-service_internal_infrastructure_entrypoints_handlers_card_dto.CardResponse'
        "400":
          description: Invalid request data
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Account not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a new card
      tags:
      - Cards
  /api/accounts/{id}/cards/{cardId}:
    delete:
      consumes:
      - application/json
      description: Delete a card
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: string
      - description: Card ID
        in: path
        name: cardId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Card deleted successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid card ID
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Card not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete card
      tags:
      - Cards
    get:
      consumes:
      - application/json
      description: Retrieve a specific card by its ID
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: string
      - description: Card ID
        in: path
        name: cardId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Card retrieved successfully
          schema:
            $ref: '#/definitions/github_com_fintrack_account-service_internal_infrastructure_entrypoints_handlers_card_dto.CardResponse'
        "400":
          description: Invalid card ID
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Card not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get card by ID
      tags:
      - Cards
    put:
      consumes:
      - application/json
      description: Update card information
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: string
      - description: Card ID
        in: path
        name: cardId
        required: true
        type: string
      - description: Card update data
        in: body
        name: card
        required: true
        schema:
          $ref: '#/definitions/github_com_fintrack_account-service_internal_infrastructure_entrypoints_handlers_card_dto.UpdateCardRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Card updated successfully
          schema:
            $ref: '#/definitions/github_com_fintrack_account-service_internal_infrastructure_entrypoints_handlers_card_dto.CardResponse'
        "400":
          description: Invalid request data
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Card not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update card
      tags:
      - Cards
  /api/accounts/{id}/cards/{cardId}/block:
    put:
      consumes:
      - application/json
      description: Block a card to prevent its usage
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: string
      - description: Card ID
        in: path
        name: cardId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Card blocked successfully
          schema:
            $ref: '#/definitions/github_com_fintrack_account-service_internal_infrastructure_entrypoints_handlers_card_dto.CardResponse'
        "400":
          description: Invalid card ID
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Card not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Block card
      tags:
      - Cards
  /api/accounts/{id}/cards/{cardId}/set-default:
    put:
      consumes:
      - application/json
      description: Set a specific card as the default for its account
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: string
      - description: Card ID
        in: path
        name: cardId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Card set as default successfully
          schema:
            $ref: '#/definitions/github_com_fintrack_account-service_internal_infrastructure_entrypoints_handlers_card_dto.CardResponse'
        "400":
          description: Invalid card ID
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Card not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Set default card
      tags:
      - Cards
  /api/accounts/{id}/cards/{cardId}/unblock:
    put:
      consumes:
      - application/json
      description: Unblock a card to allow its usage
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: string
      - description: Card ID
        in: path
        name: cardId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Card unblocked successfully
          schema:
            $ref: '#/definitions/github_com_fintrack_account-service_internal_infrastructure_entrypoints_handlers_card_dto.CardResponse'
        "400":
          description: Invalid card ID
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Card not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Unblock card
      tags:
      - Cards
...
  /api/accounts/user/{userId}/cards:
    get:
      consumes:
      - application/json
      description: Retrieve all cards owned by a user across all their accounts
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Page size
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Cards retrieved successfully
          schema:
            $ref: '#/definitions/github_com_fintrack_account-service_internal_infrastructure_entrypoints_handlers_card_dto.PaginatedCardResponse'
        "400":
          description: Invalid request parameters
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: User not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get cards by user
      tags:
      - Cards
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"